feat(assignment): extract duplicate code, load when content was deleted
